Update the README and add [compat]
This makes a minor changes to the README:
- removes the part about accepting nominations (no time left...)
- expands the installation documentation
- proposes that we split the SnoopCompile part into "pre-Cthulhu" and "post-Cthulhu"

This also adds a couple more packages (JET, SnoopCompile) and adds [compat] for long-term compatibility.

There does not seem to be a suitable Cthulhu release; either we'll need to make a branch or we can get a release out before Saturday. I favor the latter, if possible.
https://github.com/JuliaDebug/Cthulhu.jl/pull/160 looks delicious and would be nice to merge if it wouldn't jeopardize the existing functionality.